Implement DCE based on LVA
Dead code can be formulated in terms of live variable analysis.
First we have to compute when each variable is "live" (still needed
further on in the program). This can be done using the typical dataflow
equations (but written down in Datalog).

Dead code can then be formulated as each definition of a variable where
the variable is not immediately live in that same statement.